+++ CHANGELOG.md
@@ -314,3 +314,4 @@
 # 0.3.5
 * Define inductive `Denumerator` number datatype in `Object` module
 * Rename `Natural` datatype to `Numerator`
+* Define `Measurable` typeclass to measure datastructures